+++ Docs/v0/CHANGELOG.md
@@ -1,3 +1,9 @@
+V 0.1.8
+
+- Locked PredictableRandom to a hard set of random values
+- Cleaned PredictableRandom up a bit
+- Created a reproducible, serializable, and sharable state for PredictableRandom 
+
 V 0.1.7
 
 - Began cleaning hte project up more